+++ app/build.gradle.kts
@@ -145,6 +145,9 @@ dependencies {
     "fullImplementation"(libs.firebase.crashlytics)
     "fullImplementation"(libs.firebase.config)
     "fullImplementation"(libs.firebase.perf)
+    "fullImplementation"(libs.mlkit.language.id)
+    "fullImplementation"(libs.mlkit.translate)
+    "fullImplementation"(libs.opencc4j)
 
     implementation(libs.timber)
 }
