+++ app/src/foss/java/com/zionhuang/music/utils/TranslationHelper.kt
@@ -1,0 +1,8 @@
+package com.zionhuang.music.utils
+
+import com.zionhuang.music.db.entities.LyricsEntity
+
+object TranslationHelper {
+    suspend fun translate(lyrics: LyricsEntity): LyricsEntity = lyrics
+    suspend fun clearModels() {}
+}
