+++ app/src/main/java/com/zionhuang/music/playback/PlayerConnection.kt
@@ -1,5 +1,6 @@
 package com.zionhuang.music.playback
 
+import android.content.Context
 import androidx.media3.common.MediaItem
 import androidx.media3.common.PlaybackException
 import androidx.media3.common.Player
@@ -11,24 +12,31 @@ import androidx.media3.common.Player.REPEAT_MODE_OFF
 import androidx.media3.common.Player.REPEAT_MODE_ONE
 import androidx.media3.common.Player.STATE_ENDED
 import androidx.media3.common.Timeline
+import com.zionhuang.music.constants.TranslateLyricsKey
 import com.zionhuang.music.db.MusicDatabase
+import com.zionhuang.music.db.entities.LyricsEntity.Companion.LYRICS_NOT_FOUND
 import com.zionhuang.music.extensions.currentMetadata
 import com.zionhuang.music.extensions.getCurrentQueueIndex
 import com.zionhuang.music.extensions.getQueueWindows
 import com.zionhuang.music.extensions.metadata
 import com.zionhuang.music.playback.MusicService.MusicBinder
 import com.zionhuang.music.playback.queues.Queue
+import com.zionhuang.music.utils.TranslationHelper
+import com.zionhuang.music.utils.dataStore
 import com.zionhuang.music.utils.reportException
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
 import kotlinx.coroutines.flow.combine
+import kotlinx.coroutines.flow.distinctUntilChanged
 import kotlinx.coroutines.flow.flatMapLatest
+import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 
 @OptIn(ExperimentalCoroutinesApi::class)
 class PlayerConnection(
+    context: Context,
     binder: MusicBinder,
     val database: MusicDatabase,
     scope: CoroutineScope,
@@ -45,9 +53,26 @@ class PlayerConnection(
     val currentSong = mediaMetadata.flatMapLatest {
         database.song(it?.id)
     }
-    val currentLyrics = mediaMetadata.flatMapLatest { mediaMetadata ->
-        database.lyrics(mediaMetadata?.id)
-    }
+    val translating = MutableStateFlow(false)
+    val currentLyrics = combine(
+        context.dataStore.data.map {
+            it[TranslateLyricsKey] ?: true
+        }.distinctUntilChanged(),
+        mediaMetadata.flatMapLatest { mediaMetadata ->
+            database.lyrics(mediaMetadata?.id)
+        }
+    ) { translateEnabled, lyrics ->
+        if (!translateEnabled || lyrics == null || lyrics.lyrics == LYRICS_NOT_FOUND) return@combine lyrics
+        translating.value = true
+        try {
+            TranslationHelper.translate(lyrics)
+        } catch (e: Exception) {
+            reportException(e)
+            lyrics
+        }.also {
+            translating.value = false
+        }
+    }.stateIn(scope, SharingStarted.Lazily, null)
     val currentFormat = mediaMetadata.flatMapLatest { mediaMetadata ->
         database.format(mediaMetadata?.id)
     }
