+++ app/src/main/java/com/zionhuang/music/ui/component/Lyrics.kt
@@ -1,18 +1,34 @@
 package com.zionhuang.music.ui.component
 
-import android.app.SearchManager
-import android.content.Intent
-import androidx.compose.animation.animateContentSize
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.*
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.BoxWithConstraints
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.WindowInsetsSides
+import androidx.compose.foundation.layout.add
+import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.only
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.systemBars
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.foundation.lazy.rememberLazyListState
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.*
-import androidx.compose.runtime.*
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.alpha
@@ -20,35 +36,31 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.input.nestedscroll.NestedScrollConnection
 import androidx.compose.ui.input.nestedscroll.NestedScrollSource
 import androidx.compose.ui.input.nestedscroll.nestedScroll
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.input.TextFieldValue
 import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.text.style.TextOverflow
 import androidx.compose.ui.unit.Velocity
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import androidx.hilt.navigation.compose.hiltViewModel
-import com.zionhuang.music.LocalDatabase
+import com.zionhuang.music.BuildConfig
 import com.zionhuang.music.LocalPlayerConnection
 import com.zionhuang.music.R
 import com.zionhuang.music.constants.LyricsTextPositionKey
-import com.zionhuang.music.db.entities.LyricsEntity
+import com.zionhuang.music.constants.TranslateLyricsKey
 import com.zionhuang.music.db.entities.LyricsEntity.Companion.LYRICS_NOT_FOUND
 import com.zionhuang.music.lyrics.LyricsEntry
 import com.zionhuang.music.lyrics.LyricsEntry.Companion.HEAD_LYRICS_ENTRY
 import com.zionhuang.music.lyrics.LyricsUtils.findCurrentLineIndex
 import com.zionhuang.music.lyrics.LyricsUtils.parseLyrics
-import com.zionhuang.music.models.MediaMetadata
 import com.zionhuang.music.ui.component.shimmer.ShimmerHost
 import com.zionhuang.music.ui.component.shimmer.TextPlaceholder
+import com.zionhuang.music.ui.menu.LyricsMenu
 import com.zionhuang.music.ui.screens.settings.LyricsPosition
 import com.zionhuang.music.ui.utils.fadingEdge
 import com.zionhuang.music.utils.rememberEnumPreference
-import com.zionhuang.music.viewmodels.LyricsMenuViewModel
+import com.zionhuang.music.utils.rememberPreference
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.isActive
 import kotlin.time.Duration.Companion.seconds
@@ -63,11 +75,14 @@ fun Lyrics(
     val density = LocalDensity.current
 
     val lyricsTextPosition by rememberEnumPreference(LyricsTextPositionKey, LyricsPosition.CENTER)
+    var translationEnabled by rememberPreference(TranslateLyricsKey, false)
 
     val mediaMetadata by playerConnection.mediaMetadata.collectAsState()
+    val translating by playerConnection.translating.collectAsState()
     val lyricsEntity by playerConnection.currentLyrics.collectAsState(initial = null)
-    val lyrics = remember(lyricsEntity) {
-        lyricsEntity?.lyrics
+    val lyrics = remember(lyricsEntity, translating) {
+        if (translating) null
+        else lyricsEntity?.lyrics
     }
 
     val lines = remember(lyrics) {
@@ -162,31 +177,8 @@ fun Lyrics(
                 })
         ) {
             val displayedCurrentLineIndex = if (isSeeking) deferredCurrentLineIndex else currentLineIndex
-            itemsIndexed(
-                items = lines
-            ) { index, item ->
-                Text(
-                    text = item.text,
-                    fontSize = 20.sp,
-                    color = if (index == displayedCurrentLineIndex) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.secondary,
-                    textAlign = when (lyricsTextPosition) {
-                        LyricsPosition.LEFT -> TextAlign.Left
-                        LyricsPosition.CENTER -> TextAlign.Center
-                        LyricsPosition.RIGHT -> TextAlign.Right
-                    },
-                    fontWeight = FontWeight.Bold,
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable(enabled = isSynced) {
-                            playerConnection.player.seekTo(item.time)
-                            lastPreviewTime = 0L
-                        }
-                        .padding(horizontal = 24.dp, vertical = 8.dp)
-                        .alpha(if (!isSynced || index == displayedCurrentLineIndex) 1f else 0.5f)
-                )
-            }
 
-            if (lyrics == null) {
+            if (lyrics == null || translating) {
                 item {
                     ShimmerHost {
                         repeat(10) {
@@ -204,6 +196,30 @@ fun Lyrics(
                             }
                         }
                     }
+                }
+            } else {
+                itemsIndexed(
+                    items = lines
+                ) { index, item ->
+                    Text(
+                        text = item.text,
+                        fontSize = 20.sp,
+                        color = if (index == displayedCurrentLineIndex) MaterialTheme.colorScheme.primary else MaterialTheme.colorScheme.secondary,
+                        textAlign = when (lyricsTextPosition) {
+                            LyricsPosition.LEFT -> TextAlign.Left
+                            LyricsPosition.CENTER -> TextAlign.Center
+                            LyricsPosition.RIGHT -> TextAlign.Right
+                        },
+                        fontWeight = FontWeight.Bold,
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .clickable(enabled = isSynced) {
+                                playerConnection.player.seekTo(item.time)
+                                lastPreviewTime = 0L
+                            }
+                            .padding(horizontal = 24.dp, vertical = 8.dp)
+                            .alpha(if (!isSynced || index == displayedCurrentLineIndex) 1f else 0.5f)
+                    )
                 }
             }
         }
@@ -227,277 +243,42 @@ fun Lyrics(
         }
 
         mediaMetadata?.let { mediaMetadata ->
-            IconButton(
+            Row(
                 modifier = Modifier
                     .align(Alignment.BottomEnd)
-                    .padding(12.dp),
-                onClick = {
-                    menuState.show {
-                        LyricsMenu(
-                            lyricsProvider = { lyricsEntity },
-                            mediaMetadataProvider = { mediaMetadata },
-                            onDismiss = menuState::dismiss
-                        )
-                    }
-                }
+                    .padding(end = 12.dp)
             ) {
-                Icon(
-                    painter = painterResource(id = R.drawable.more_horiz),
-                    contentDescription = null
-                )
-            }
-        }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun LyricsMenu(
-    lyricsProvider: () -> LyricsEntity?,
-    mediaMetadataProvider: () -> MediaMetadata,
-    onDismiss: () -> Unit,
-    viewModel: LyricsMenuViewModel = hiltViewModel(),
-) {
-    val context = LocalContext.current
-    val database = LocalDatabase.current
-
-    var showEditDialog by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    if (showEditDialog) {
-        TextFieldDialog(
-            onDismiss = { showEditDialog = false },
-            icon = { Icon(painter = painterResource(R.drawable.edit), contentDescription = null) },
-            title = { Text(text = mediaMetadataProvider().title) },
-            initialTextFieldValue = TextFieldValue(lyricsProvider()?.lyrics.orEmpty()),
-            singleLine = false,
-            onDone = {
-                database.query {
-                    upsert(
-                        LyricsEntity(
-                            id = mediaMetadataProvider().id,
-                            lyrics = it
-                        )
-                    )
-                }
-            }
-        )
-    }
-
-    var showSearchDialog by rememberSaveable {
-        mutableStateOf(false)
-    }
-    var showSearchResultDialog by rememberSaveable {
-        mutableStateOf(false)
-    }
-
-    val searchMediaMetadata = remember(showSearchDialog) {
-        mediaMetadataProvider()
-    }
-    val (titleField, onTitleFieldChange) = rememberSaveable(showSearchDialog, stateSaver = TextFieldValue.Saver) {
-        mutableStateOf(
-            TextFieldValue(
-                text = mediaMetadataProvider().title
-            )
-        )
-    }
-    val (artistField, onArtistFieldChange) = rememberSaveable(showSearchDialog, stateSaver = TextFieldValue.Saver) {
-        mutableStateOf(
-            TextFieldValue(
-                text = mediaMetadataProvider().artists.joinToString { it.name }
-            )
-        )
-    }
-
-    if (showSearchDialog) {
-        DefaultDialog(
-            modifier = Modifier.verticalScroll(rememberScrollState()),
-            onDismiss = { showSearchDialog = false },
-            icon = { Icon(painter = painterResource(R.drawable.search), contentDescription = null) },
-            title = { Text(stringResource(R.string.search_lyrics)) },
-            buttons = {
-                TextButton(
-                    onClick = { showSearchDialog = false }
-                ) {
-                    Text(stringResource(android.R.string.cancel))
-                }
-
-                Spacer(Modifier.width(8.dp))
-
-                TextButton(
-                    onClick = {
-                        showSearchDialog = false
-                        onDismiss()
-                        try {
-                            context.startActivity(
-                                Intent(Intent.ACTION_WEB_SEARCH).apply {
-                                    putExtra(SearchManager.QUERY, "${artistField.text} ${titleField.text} lyrics")
-                                }
-                            )
-                        } catch (_: Exception) {
-                        }
-                    }
-                ) {
-                    Text(stringResource(R.string.search_online))
-                }
-
-                Spacer(Modifier.width(8.dp))
-
-                TextButton(
-                    onClick = {
-                        viewModel.search(searchMediaMetadata.id, titleField.text, artistField.text, searchMediaMetadata.duration)
-                        showSearchResultDialog = true
-                    }
-                ) {
-                    Text(stringResource(android.R.string.ok))
-                }
-            }
-        ) {
-            OutlinedTextField(
-                value = titleField,
-                onValueChange = onTitleFieldChange,
-                singleLine = true,
-                label = { Text(stringResource(R.string.song_title)) }
-            )
-
-            Spacer(Modifier.height(12.dp))
-
-            OutlinedTextField(
-                value = artistField,
-                onValueChange = onArtistFieldChange,
-                singleLine = true,
-                label = { Text(stringResource(R.string.song_artists)) }
-            )
-        }
-    }
-
-    if (showSearchResultDialog) {
-        val results by viewModel.results.collectAsState()
-        val isLoading by viewModel.isLoading.collectAsState()
-
-        var expandedItemIndex by rememberSaveable {
-            mutableStateOf(-1)
-        }
-
-        ListDialog(
-            onDismiss = { showSearchResultDialog = false }
-        ) {
-            itemsIndexed(results) { index, result ->
-                Row(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .clickable {
-                            onDismiss()
-                            viewModel.cancelSearch()
-                            database.query {
-                                upsert(
-                                    LyricsEntity(
-                                        id = searchMediaMetadata.id,
-                                        lyrics = result.lyrics
-                                    )
-                                )
-                            }
-                        }
-                        .padding(12.dp)
-                        .animateContentSize()
-                ) {
-                    Column(
-                        modifier = Modifier.weight(1f)
-                    ) {
-                        Text(
-                            text = result.lyrics,
-                            style = MaterialTheme.typography.bodyMedium,
-                            maxLines = if (index == expandedItemIndex) Int.MAX_VALUE else 2,
-                            overflow = TextOverflow.Ellipsis,
-                            modifier = Modifier.padding(bottom = 4.dp)
-                        )
-
-                        Row(
-                            verticalAlignment = Alignment.CenterVertically
-                        ) {
-                            Text(
-                                text = result.providerName,
-                                style = MaterialTheme.typography.labelMedium,
-                                color = MaterialTheme.colorScheme.secondary,
-                                maxLines = 1
-                            )
-                            if (result.lyrics.startsWith("[")) {
-                                Icon(
-                                    painter = painterResource(R.drawable.sync),
-                                    contentDescription = null,
-                                    tint = MaterialTheme.colorScheme.secondary,
-                                    modifier = Modifier
-                                        .padding(start = 4.dp)
-                                        .size(18.dp)
-                                )
-                            }
-                        }
-                    }
-
+                if (BuildConfig.FLAVOR == "full") {
                     IconButton(
                         onClick = {
-                            expandedItemIndex = if (expandedItemIndex == index) -1 else index
+                            translationEnabled = !translationEnabled
                         }
                     ) {
                         Icon(
-                            painter = painterResource(if (index == expandedItemIndex) R.drawable.expand_less else R.drawable.expand_more),
-                            contentDescription = null
+                            painter = painterResource(id = R.drawable.translate),
+                            contentDescription = null,
+                            tint = LocalContentColor.current.copy(alpha = if (translationEnabled) 1f else 0.3f)
                         )
                     }
                 }
-            }
 
-            if (isLoading) {
-                item {
-                    Box(
-                        contentAlignment = Alignment.Center,
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        CircularProgressIndicator()
+                IconButton(
+                    onClick = {
+                        menuState.show {
+                            LyricsMenu(
+                                lyricsProvider = { lyricsEntity },
+                                mediaMetadataProvider = { mediaMetadata },
+                                onDismiss = menuState::dismiss
+                            )
+                        }
                     }
-                }
-            }
-
-            if (!isLoading && results.isEmpty()) {
-                item {
-                    Text(
-                        text = context.getString(R.string.lyrics_not_found),
-                        textAlign = TextAlign.Center,
-                        modifier = Modifier
-                            .fillMaxWidth()
+                ) {
+                    Icon(
+                        painter = painterResource(id = R.drawable.more_horiz),
+                        contentDescription = null
                     )
                 }
             }
-        }
-    }
-
-    GridMenu(
-        contentPadding = PaddingValues(
-            start = 8.dp,
-            top = 8.dp,
-            end = 8.dp,
-            bottom = 8.dp + WindowInsets.systemBars.asPaddingValues().calculateBottomPadding()
-        )
-    ) {
-        GridMenuItem(
-            icon = R.drawable.edit,
-            title = R.string.edit
-        ) {
-            showEditDialog = true
-        }
-        GridMenuItem(
-            icon = R.drawable.cached,
-            title = R.string.refetch
-        ) {
-            onDismiss()
-            viewModel.refetchLyrics(mediaMetadataProvider(), lyricsProvider())
-        }
-        GridMenuItem(
-            icon = R.drawable.search,
-            title = R.string.search,
-        ) {
-            showSearchDialog = true
         }
     }
 }
