+++ app/src/main/java/com/zionhuang/music/ui/menu/LyricsMenu.kt
@@ -1,0 +1,310 @@
+package com.zionhuang.music.ui.menu
+
+import android.app.SearchManager
+import android.content.Intent
+import androidx.compose.animation.animateContentSize
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.WindowInsets
+import androidx.compose.foundation.layout.asPaddingValues
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.systemBars
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.input.TextFieldValue
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.unit.dp
+import androidx.hilt.navigation.compose.hiltViewModel
+import com.zionhuang.music.LocalDatabase
+import com.zionhuang.music.R
+import com.zionhuang.music.db.entities.LyricsEntity
+import com.zionhuang.music.models.MediaMetadata
+import com.zionhuang.music.ui.component.DefaultDialog
+import com.zionhuang.music.ui.component.GridMenu
+import com.zionhuang.music.ui.component.GridMenuItem
+import com.zionhuang.music.ui.component.ListDialog
+import com.zionhuang.music.ui.component.TextFieldDialog
+import com.zionhuang.music.viewmodels.LyricsMenuViewModel
+
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+fun LyricsMenu(
+    lyricsProvider: () -> LyricsEntity?,
+    mediaMetadataProvider: () -> MediaMetadata,
+    onDismiss: () -> Unit,
+    viewModel: LyricsMenuViewModel = hiltViewModel(),
+) {
+    val context = LocalContext.current
+    val database = LocalDatabase.current
+
+    var showEditDialog by rememberSaveable {
+        mutableStateOf(false)
+    }
+
+    if (showEditDialog) {
+        TextFieldDialog(
+            onDismiss = { showEditDialog = false },
+            icon = { Icon(painter = painterResource(R.drawable.edit), contentDescription = null) },
+            title = { Text(text = mediaMetadataProvider().title) },
+            initialTextFieldValue = TextFieldValue(lyricsProvider()?.lyrics.orEmpty()),
+            singleLine = false,
+            onDone = {
+                database.query {
+                    upsert(
+                        LyricsEntity(
+                            id = mediaMetadataProvider().id,
+                            lyrics = it
+                        )
+                    )
+                }
+            }
+        )
+    }
+
+    var showSearchDialog by rememberSaveable {
+        mutableStateOf(false)
+    }
+    var showSearchResultDialog by rememberSaveable {
+        mutableStateOf(false)
+    }
+
+    val searchMediaMetadata = remember(showSearchDialog) {
+        mediaMetadataProvider()
+    }
+    val (titleField, onTitleFieldChange) = rememberSaveable(showSearchDialog, stateSaver = TextFieldValue.Saver) {
+        mutableStateOf(
+            TextFieldValue(
+                text = mediaMetadataProvider().title
+            )
+        )
+    }
+    val (artistField, onArtistFieldChange) = rememberSaveable(showSearchDialog, stateSaver = TextFieldValue.Saver) {
+        mutableStateOf(
+            TextFieldValue(
+                text = mediaMetadataProvider().artists.joinToString { it.name }
+            )
+        )
+    }
+
+    if (showSearchDialog) {
+        DefaultDialog(
+            modifier = Modifier.verticalScroll(rememberScrollState()),
+            onDismiss = { showSearchDialog = false },
+            icon = { Icon(painter = painterResource(R.drawable.search), contentDescription = null) },
+            title = { Text(stringResource(R.string.search_lyrics)) },
+            buttons = {
+                TextButton(
+                    onClick = { showSearchDialog = false }
+                ) {
+                    Text(stringResource(android.R.string.cancel))
+                }
+
+                Spacer(Modifier.width(8.dp))
+
+                TextButton(
+                    onClick = {
+                        showSearchDialog = false
+                        onDismiss()
+                        try {
+                            context.startActivity(
+                                Intent(Intent.ACTION_WEB_SEARCH).apply {
+                                    putExtra(SearchManager.QUERY, "${artistField.text} ${titleField.text} lyrics")
+                                }
+                            )
+                        } catch (_: Exception) {
+                        }
+                    }
+                ) {
+                    Text(stringResource(R.string.search_online))
+                }
+
+                Spacer(Modifier.width(8.dp))
+
+                TextButton(
+                    onClick = {
+                        viewModel.search(searchMediaMetadata.id, titleField.text, artistField.text, searchMediaMetadata.duration)
+                        showSearchResultDialog = true
+                    }
+                ) {
+                    Text(stringResource(android.R.string.ok))
+                }
+            }
+        ) {
+            OutlinedTextField(
+                value = titleField,
+                onValueChange = onTitleFieldChange,
+                singleLine = true,
+                label = { Text(stringResource(R.string.song_title)) }
+            )
+
+            Spacer(Modifier.height(12.dp))
+
+            OutlinedTextField(
+                value = artistField,
+                onValueChange = onArtistFieldChange,
+                singleLine = true,
+                label = { Text(stringResource(R.string.song_artists)) }
+            )
+        }
+    }
+
+    if (showSearchResultDialog) {
+        val results by viewModel.results.collectAsState()
+        val isLoading by viewModel.isLoading.collectAsState()
+
+        var expandedItemIndex by rememberSaveable {
+            mutableStateOf(-1)
+        }
+
+        ListDialog(
+            onDismiss = { showSearchResultDialog = false }
+        ) {
+            itemsIndexed(results) { index, result ->
+                Row(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .clickable {
+                            onDismiss()
+                            viewModel.cancelSearch()
+                            database.query {
+                                upsert(
+                                    LyricsEntity(
+                                        id = searchMediaMetadata.id,
+                                        lyrics = result.lyrics
+                                    )
+                                )
+                            }
+                        }
+                        .padding(12.dp)
+                        .animateContentSize()
+                ) {
+                    Column(
+                        modifier = Modifier.weight(1f)
+                    ) {
+                        Text(
+                            text = result.lyrics,
+                            style = MaterialTheme.typography.bodyMedium,
+                            maxLines = if (index == expandedItemIndex) Int.MAX_VALUE else 2,
+                            overflow = TextOverflow.Ellipsis,
+                            modifier = Modifier.padding(bottom = 4.dp)
+                        )
+
+                        Row(
+                            verticalAlignment = Alignment.CenterVertically
+                        ) {
+                            Text(
+                                text = result.providerName,
+                                style = MaterialTheme.typography.labelMedium,
+                                color = MaterialTheme.colorScheme.secondary,
+                                maxLines = 1
+                            )
+                            if (result.lyrics.startsWith("[")) {
+                                Icon(
+                                    painter = painterResource(R.drawable.sync),
+                                    contentDescription = null,
+                                    tint = MaterialTheme.colorScheme.secondary,
+                                    modifier = Modifier
+                                        .padding(start = 4.dp)
+                                        .size(18.dp)
+                                )
+                            }
+                        }
+                    }
+
+                    IconButton(
+                        onClick = {
+                            expandedItemIndex = if (expandedItemIndex == index) -1 else index
+                        }
+                    ) {
+                        Icon(
+                            painter = painterResource(if (index == expandedItemIndex) R.drawable.expand_less else R.drawable.expand_more),
+                            contentDescription = null
+                        )
+                    }
+                }
+            }
+
+            if (isLoading) {
+                item {
+                    Box(
+                        contentAlignment = Alignment.Center,
+                        modifier = Modifier.fillMaxWidth()
+                    ) {
+                        CircularProgressIndicator()
+                    }
+                }
+            }
+
+            if (!isLoading && results.isEmpty()) {
+                item {
+                    Text(
+                        text = context.getString(R.string.lyrics_not_found),
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier
+                            .fillMaxWidth()
+                    )
+                }
+            }
+        }
+    }
+
+    GridMenu(
+        contentPadding = PaddingValues(
+            start = 8.dp,
+            top = 8.dp,
+            end = 8.dp,
+            bottom = 8.dp + WindowInsets.systemBars.asPaddingValues().calculateBottomPadding()
+        )
+    ) {
+        GridMenuItem(
+            icon = R.drawable.edit,
+            title = R.string.edit
+        ) {
+            showEditDialog = true
+        }
+        GridMenuItem(
+            icon = R.drawable.cached,
+            title = R.string.refetch
+        ) {
+            onDismiss()
+            viewModel.refetchLyrics(mediaMetadataProvider(), lyricsProvider())
+        }
+        GridMenuItem(
+            icon = R.drawable.search,
+            title = R.string.search,
+        ) {
+            showSearchDialog = true
+        }
+    }
+}
