+++ app/src/main/java/com/zionhuang/music/ui/menu/PlayerMenu.kt
@@ -277,7 +277,7 @@ fun PlayerMenu(
         }
         GridMenuItem(
             icon = R.drawable.tune,
-            title = R.string.playback_control
+            title = R.string.advanced
         ) {
             showPitchTempoDialog = true
         }
