+++ app/src/main/res/values-es/strings.xml
@@ -87,7 +87,7 @@
     <string name="remove_from_history">Eliminar del historial</string>
     <string name="search_online">Buscar online</string>
     <string name="sync">Sincronizar</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Fecha añadida</string>
@@ -269,4 +269,6 @@
     <string name="app_version">Versión de la app</string>
 
     <string name="new_version_available">New version available</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
