+++ app/src/main/res/values-fa-rIR/strings.xml
@@ -86,7 +86,7 @@
     <string name="remove_from_history">Remove from history</string>
     <string name="search_online">Search online</string>
     <string name="sync">Sync</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">تاریخ اضافه‌شده</string>
@@ -260,4 +260,6 @@
     <string name="app_version">نسخه‌ی برنامه</string>
 
     <string name="new_version_available">New version available</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
