+++ app/src/main/res/values-fr-rFR/strings.xml
@@ -87,7 +87,7 @@
     <string name="remove_from_history">Supprimer de l\'historique</string>
     <string name="search_online">Rechercher en ligne</string>
     <string name="sync">Synchroniser</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Date d\'ajout</string>
@@ -269,4 +269,6 @@
     <string name="app_version">Version de l\'application</string>
 
     <string name="new_version_available">Nouvelle version disponible</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
