+++ app/src/main/res/values-hu/strings.xml
@@ -86,7 +86,7 @@
     <string name="remove_from_history">Előzményből eltávolít</string>
     <string name="search_online">Keresés online</string>
     <string name="sync">Szinkron.</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Létrehozás dátuma</string>
@@ -260,4 +260,6 @@
     <string name="app_version">App verzió</string>
 
     <string name="new_version_available">New version available</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
