+++ app/src/main/res/values-id/strings.xml
@@ -85,7 +85,7 @@
     <string name="remove_from_history">Remove from history</string>
     <string name="search_online">Cari secara online</string>
     <string name="sync">Sync</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Tanggal ditambahkan</string>
@@ -251,4 +251,6 @@
     <string name="app_version">Versi aplikasi</string>
 
     <string name="new_version_available">New version available</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
