+++ app/src/main/res/values-it/strings.xml
@@ -87,7 +87,7 @@
     <string name="remove_from_history">Rimuovi da cronologia</string>
     <string name="search_online">Cerca online</string>
     <string name="sync">Sincronizza</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Data di aggiunta</string>
@@ -269,4 +269,6 @@
     <string name="app_version">Versione dell\'app</string>
 
     <string name="new_version_available">New version available</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
