+++ app/src/main/res/values-pl/strings.xml
@@ -88,7 +88,7 @@
     <string name="remove_from_history">Usuń z historii</string>
     <string name="search_online">Szukaj online</string>
     <string name="sync">Synchronizuj</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">Advanced</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">Data dodania</string>
@@ -278,4 +278,6 @@
     <string name="app_version">Wersja aplikacji</string>
 
     <string name="new_version_available">Dostępna nowa wersja</string>
+    <string name="translation_models">Translation Models</string>
+    <string name="clear_translation_models">Clear translation models</string>
 </resources>
