+++ app/src/main/res/values-zh-rTW/strings.xml
@@ -61,7 +61,7 @@
     <string name="retry">重試</string>
     <string name="radio">電台</string>
     <string name="shuffle">隨機播放</string>
-    <string name="reset">Reset</string>
+    <string name="reset">重設</string>
 
     <!-- Menu -->
     <string name="details">詳細資訊</string>
@@ -85,7 +85,7 @@
     <string name="remove_from_history">從記錄中移除</string>
     <string name="search_online">線上搜尋</string>
     <string name="sync">同步</string>
-    <string name="playback_control">Playback control</string>
+    <string name="advanced">進階</string>
 
     <!-- Sort menu -->
     <string name="sort_by_create_date">新增時間</string>
@@ -251,4 +251,6 @@
     <string name="app_version">應用程式版本</string>
 
     <string name="new_version_available">發現新版本</string>
+    <string name="translation_models">翻譯模型</string>
+    <string name="clear_translation_models">清除翻譯模型</string>
 </resources>
